Architecture: Allow overriding memory allocation functions

+++ tests/testing-plugins/testing_clock.c
@@ -42,7 +42,7 @@ UA_realSleep(UA_UInt32 duration) {
 }
 
 void
-UA_comboSleep(size_t duration) {
+UA_comboSleep(unsigned long duration) {
     UA_fakeSleep((UA_UInt32)duration);
     UA_realSleep((UA_UInt32)duration);
 }
